requirements.txt
flask==3.0.0
flask-cors==4.0.0
edgartools>=4.0.0
pandas>=2.1.0
python-dotenv>=1.0.0
alpaca-py>=0.42.0
requests>=2.31.0
supabase>=2.0.0
